backend/pyproject.toml
[project]
name = "recipe-companion-backend"
version = "0.1.0"
description = "Recipe analysis agent with CopilotKit integration"
requires-python = ">=3.11"
dependencies = [
    "copilotkit>=0.1.87",
    "pydantic-ai>=1.86.0",
    "ag-ui-protocol>=0.1.18",
    "fastapi>=0.136.0",
    "uvicorn>=0.46.0",
    "python-multipart>=0.0.26",
    "pypdf>=6.10.2",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "starlette<1.0",
    "warp-cache>=0.7.0",
]

[project.optional-dependencies]
test = [
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]
dev = [
    "ipdb>=0.13.13",
    "ruff>=0.15.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests that require real API calls (deselect with '-m \"not integration\"')",
]
